LayoutTests/fast/dom/HTMLElement/insertAdjacentHTML-errors-expected.txt

Test insertAdjacentHTML exceptions to make sure they match HTML5

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS div.insertAdjacentHTML('beforeBegin', 'text') threw exception NoModificationAllowedError (DOM Exception 7): The object can not be modified..
PASS div.insertAdjacentHTML('afterEnd', 'text') threw exception NoModificationAllowedError (DOM Exception 7): The object can not be modified..
PASS div.insertAdjacentHTML('FOO', 'text') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.documentElement.insertAdjacentHTML('afterEnd', 'text') threw exception NoModificationAllowedError (DOM Exception 7): The object can not be modified..
PASS successfullyParsed is true

TEST COMPLETE
